#74 - refactor : 학습 기록 기능에서 변경된 시큐리티 설정 적용.
