+++ src/main/java/com/example/masterplanbbe/domain/studyLog/entity/StudyLog.java
@@ -2,9 +2,9 @@ package com.example.masterplanbbe.domain.studyLog.entity;
 
 import com.example.masterplanbbe.common.domain.FullAuditEntity;
 import com.example.masterplanbbe.domain.exam.entity.Exam;
+import com.example.masterplanbbe.domain.member.entity.Member;
 import com.example.masterplanbbe.domain.studyLog.enums.InputSource;
 import com.example.masterplanbbe.domain.studyLog.request.StudyLogRequest;
-import com.example.masterplanbbe.member.entity.Member;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
